split profile and profileStats
